Big Update

+++ drago.js
@@ -6,6 +6,7 @@ const mongo = require('./mongo')
 const splitz = require('./models/split-schema')
 const person = require('./models/person')
 const regear = require('./models/regear')
+const report = require('./models/report')
 const register = require('./models/register')
 const axios = require('axios'); 
 const prefix = '!'
@@ -542,15 +543,75 @@ client.on('message', async (message) => {
          }
     } else if (command === 'handle') {
         if (message.member.roles.cache.has('759793776439984170')) {
-            if (message.channel.parentID === '853522303811321876') {
-                const ticketNumber = message.channel.name.slice('ticket'.length);
-                const channel = message.guild.channels.cache.get('779514684797091850');
-                const channelID = message.channel.id;
-                message.channel.setName(`gremory${ticketNumber}`)
-                channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
-                message.delete()
-                console.log(ticketNumber)
-            }
+            await mongo().then(async mongoose => {
+                if (message.channel.parentID === '853522303811321876') {
+                    let isHandled = await report.findOne({ officerId: message.author.id})
+                    if (isHandled) {
+                        return message.author.send(`I'm sorry but ${message.guild.members.cache.get(isHandled.officerId).nickname} is handling this case`)
+                    } else {
+                        if (message.author.id === '694488949980135444') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`gremory${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '382507136883621889') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`luai${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '412598799081406465') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`amorkia${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '229238295257677835') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`fatmeow${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '590791325759045643') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`gnamo${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '409717155035217922') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`lightdragneel${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        } else if (message.author.id === '263996887131095041') {
+                            const ticketNumber = message.channel.name.slice('ticket'.length);
+                            const channel = message.guild.channels.cache.get('779514684797091850');
+                            const channelID = message.channel.id;
+                            message.channel.setName(`lightdragneel${ticketNumber}`)
+                            channel.send(`<@${message.author.id}> is handling case <#${channelID}>`)
+                            message.delete()
+                            console.log(ticketNumber)
+                        }
+                    }
+                } else {
+                    message.delete()
+                    return message.author.send('This command is used to handle cases!')
+                }
+            })
         } else {
             message.delete()
             return message.author.send('I\'m sorry but you don\'t have the right to use this command')
